feat: consolidated Threads.java into the StudentQAPage.java application
