+++ FoundationCode/src/application/StudentQAPage.java
@@ -1,6 +1,7 @@
 package application;
 
 import databasePart1.DatabaseHelper;
+import databasePart1.DatabaseHelper.PrivateMessage;
 import javafx.geometry.Insets;
 import javafx.geometry.Pos;
 import javafx.scene.Scene;
@@ -11,9 +12,12 @@ import javafx.scene.control.Alert.AlertType;
 import javafx.scene.image.Image;
 
 import java.sql.SQLException;
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+import java.util.ArrayList;
+import java.util.Comparator;
 import java.util.List;
 import java.util.Optional;
-
 
 /**
 
@@ -29,7 +33,14 @@ public class StudentQAPage {
     private Stage primaryStage;
     private VBox mainLayout;
     private TabPane tabPane;
+    private VBox displayQuestions;
+    private VBox displayQuestionThread;
+    private ComboBox<String> viewFilter; // Public / Private / All
+    private ComboBox<String> resolutionFilter;   // Resolved / Unresolved / All
+    private Label allQuestionsError;             // Inline error banner for the All Questions tab
 	
+    private static final DateTimeFormatter TS = DateTimeFormatter.ofPattern("MMM dd, yyyy HH:mm");
+    
     // Tabs
     private Tab myQuestionsTab;
     private Tab allQuestionsTab;
@@ -572,68 +583,717 @@ public class StudentQAPage {
     private Tab createAllQuestionsTab() {
         Tab tab = new Tab("All Questions");
         tab.setClosable(false);
+        
         VBox content = new VBox(10);
         content.setPadding(new Insets(20));
         content.setStyle("-fx-background-color: white;");
 
-		
-        // Filter controls
-        HBox filterBox = new HBox(10);
-        filterBox.setAlignment(Pos.CENTER_LEFT);
-        Label filterLabel = new Label("Show:");
-        filterLabel.setStyle("-fx-text-fill: #000;");
-        ComboBox<String> filterCombo = new ComboBox<>();
-        filterCombo.getItems().addAll("All Questions", "Unresolved Only", "Resolved Only");
-        filterCombo.setValue("Unresolved Only");
-        
-        filterBox.getChildren().addAll(filterLabel, filterCombo);
-        
-        VBox questionsContainer = new VBox(10);
+        Label signedIn = new Label("Signed in as: " + currentUser.getUserName() + " (" + currentUser.getRole() + ")");
+        signedIn.setStyle("-fx-text-fill:#444; -fx-font-size:12px;");
 
-		
-        // Load questions based on filter
-        filterCombo.setOnAction(e -> {
+        viewFilter = new ComboBox<>();
+        viewFilter.getItems().addAll("Public", "Private", "All");
+        viewFilter.setValue("All"); // default
+        viewFilter.setTooltip(new Tooltip("Choose which posts to show in the thread"));
+
+        HBox topBar = new HBox(12, new Separator(), signedIn, new Label("View:") {{ setStyle("-fx-text-fill: black;"); }}, viewFilter);
+        topBar.setAlignment(Pos.CENTER_LEFT);
+        topBar.setPadding(new Insets(8, 10, 8, 10));
+        topBar.setStyle("-fx-background-color:#f3f3f3;");
+
+        // --- Left panel (questions list + New button) ---
+        displayQuestions = new VBox(8);
+        displayQuestions.setPadding(new Insets(10));
+
+        Button newQuestionBtn = new Button("New Question");
+        newQuestionBtn.setOnAction(e -> openCreateQuestionDialog());
+
+        // Resolution filter (matches Threads-style simple controls)
+        resolutionFilter = new ComboBox<>();
+        resolutionFilter.getItems().addAll("All Questions", "Unresolved Only", "Resolved Only");
+        resolutionFilter.setValue("All Questions"); // default
+        resolutionFilter.setTooltip(new Tooltip("Filter by resolution status"));
+
+        Label resolutionLbl = new Label("Resolution:");
+        resolutionLbl.setStyle("-fx-text-fill: black; -fx-font-size:12px;");
+
+        HBox filterBox = new HBox(8, resolutionLbl, resolutionFilter);
+        filterBox.setAlignment(Pos.CENTER_LEFT);
+
+        // Inline error banner (kept subtle; only shown on failures)
+        allQuestionsError = new Label();
+        allQuestionsError.setStyle("-fx-text-fill: black; -fx-font-size:11px;");
+        allQuestionsError.setVisible(false);
+
+        // Rebuild the left side to include controls + error + list
+        VBox leftSide = new VBox(10, newQuestionBtn, filterBox, allQuestionsError, displayQuestions);
+        leftSide.setPadding(new Insets(10));
+
+        ScrollPane scrollLeft = new ScrollPane(leftSide);
+        scrollLeft.setFitToWidth(true);
+
+        // --- Right panel (thread content) ---
+        displayQuestionThread = new VBox(10);
+        displayQuestionThread.setPadding(new Insets(12));
+
+        ScrollPane scrollRight = new ScrollPane(displayQuestionThread);
+        scrollRight.setFitToWidth(true);
+
+        SplitPane split = new SplitPane(scrollLeft, scrollRight);
+        split.setDividerPositions(0.35);
+
+        BorderPane root = new BorderPane();
+        root.setTop(topBar);
+        root.setCenter(split);
+        
+        // Change handler for the resolution filter
+        resolutionFilter.setOnAction(e -> {
+            allQuestionsError.setVisible(false);
             try {
-                String filter = filterCombo.getValue();
+                String f = resolutionFilter.getValue();
                 List<Question> questions;
-                
-                if ("Unresolved Only".equals(filter)) {
+
+                if ("Unresolved Only".equals(f)) {
                     questions = databaseHelper.getUnresolvedQuestions();
-                } else if ("Resolved Only".equals(filter)) {
+                } else if ("Resolved Only".equals(f)) {
                     questions = databaseHelper.getAllQuestions(null);
                     questions.removeIf(q -> !q.isResolved());
                 } else {
-                    questions = databaseHelper.getAllQuestions(null);
+                    // default to top-level thread list ordering
+                    questions = databaseHelper.getQuestionsTopLevel();
                 }
-                
-                displayAllQuestions(questionsContainer, questions);
+
+                // Re-render the left-panel question list (Threads-style rows)
+                displayQuestions.getChildren().clear();
+                if (questions.isEmpty()) {
+                    showEmptyLeftState();
+                    showEmptyRightState();
+                } else {
+                    for (Question q : questions) {
+                        displayQuestions.getChildren().add(questionRow(q));
+                    }
+                    // Keep UX parity: select the first row automatically
+                    loadThread(questions.get(0));
+                }
             } catch (SQLException ex) {
-                questionsContainer.getChildren().clear();
-                Label errorLabel = new Label("Error loading questions: " + ex.getMessage());
-                errorLabel.setStyle("-fx-text-fill: red;");
-                questionsContainer.getChildren().add(errorLabel);
-                ex.printStackTrace();
+                displayQuestions.getChildren().clear();
+                allQuestionsError.setText("Error loading questions: " + ex.getMessage());
+                allQuestionsError.setVisible(true);
             }
         });
-        
-        // Initial load
-        try {
-            List<Question> questions = databaseHelper.getUnresolvedQuestions();
-            displayAllQuestions(questionsContainer, questions);
-        } catch (SQLException e) {
-            Label errorLabel = new Label("Error loading questions: " + e.getMessage());
-            errorLabel.setStyle("-fx-text-fill: red;");
-            questionsContainer.getChildren().add(errorLabel);
-            e.printStackTrace();
-        }
-        
-        content.getChildren().addAll(filterBox, new Separator(), questionsContainer);
+
+        // Initial load honoring the default filter
+        resolutionFilter.getOnAction().handle(null);
+
+        content.getChildren().addAll(root, new Separator(), split);
         
         ScrollPane scrollPane = new ScrollPane(content);
         scrollPane.setFitToWidth(true);
         tab.setContent(scrollPane);
         
+        root.setStyle("-fx-text-fill: black; -fx-text-inner-color: black;");
+        
         return tab;
+    }
+    
+    private void refreshQuestionListAndMaybeSelect() {
+        displayQuestions.getChildren().clear();
+        try {
+            List<Question> questions = databaseHelper.getQuestionsTopLevel();
+            if (questions.isEmpty()) {
+                showEmptyLeftState();
+                showEmptyRightState();
+                return;
+            }
+            for (Question q : questions) {
+                displayQuestions.getChildren().add(questionRow(q));
+            }
+            // Select the first question by default
+            loadThread(questions.get(0));
+        } catch (SQLException ex) {
+            showError("Load failed", ex.getMessage());
+        }
+    }
+
+    private void showEmptyLeftState() {
+        Label none = new Label("No questions yet.");
+        none.setStyle("-fx-text-fill: black;");
+        VBox box = new VBox(6, none);
+        box.setPadding(new Insets(10));
+        displayQuestions.getChildren().add(box);
+    }
+
+    private void showEmptyRightState() {
+        displayQuestionThread.getChildren().clear();
+        Label title = new Label("Start the discussion");
+        title.setStyle("-fx-font-size:18px; -fx-font-weight:bold;");
+        Label text = new Label("There are no questions yet. Create one to begin.");
+        Button create = new Button("Create your first question");
+        create.setOnAction(e -> openCreateQuestionDialog());
+        VBox inner = new VBox(10, title, text, create);
+        inner.setAlignment(Pos.CENTER);
+        inner.setPadding(new Insets(30));
+        displayQuestionThread.getChildren().add(inner);
+    }
+
+    private HBox questionRow(Question q) {
+        // Title
+        Label title = new Label(q.getTitle());
+        title.setStyle("-fx-font-weight:bold; -fx-text-fill: black;");
+        
+        // First metadata line: author and date
+        Label meta1 = new Label("by " + q.getAskedBy() + " • " + q.getCreatedAt().format(TS));
+        meta1.setStyle("-fx-text-fill: black; -fx-font-size:11px;");
+        
+        // Second metadata line: answer counts and resolution status
+        StringBuilder meta2Text = new StringBuilder();
+        
+        try {
+            int publicAnswerCount = databaseHelper.getAnswersForQuestion(q.getId()).size();
+            List<PrivateMessage> privateMessages = databaseHelper.getPrivateMessagesForQuestion(q.getId());
+            
+            // Count visible private messages
+            int visiblePrivateCount = 0;
+            for (PrivateMessage pm : privateMessages) {
+                if (canSeePrivate(q, pm)) {
+                    visiblePrivateCount++;
+                }
+            }
+            
+            // Display answer counts
+            if (visiblePrivateCount > 0) {
+                meta2Text.append(publicAnswerCount).append(" public + ")
+                         .append(visiblePrivateCount).append(" private");
+            } else {
+                meta2Text.append(publicAnswerCount).append(" answer");
+                if (publicAnswerCount != 1) meta2Text.append("s");
+            }
+        } catch (SQLException ex) {
+            meta2Text.append("answers unavailable");
+        }
+        
+        // Add resolution status
+        if (q.isResolved()) {
+            meta2Text.append(" • ✓ Resolved");
+        }
+        
+        Label meta2 = new Label(meta2Text.toString());
+        meta2.setStyle("-fx-text-fill: #666; -fx-font-size:10px;");
+        
+        // Stack all labels vertically
+        VBox col = new VBox(2, title, meta1, meta2);
+        col.setPadding(new Insets(8));
+
+        // Make whole row clickable
+        HBox row = new HBox(col);
+        row.setStyle("-fx-background-color:#ffffff; -fx-border-color:#e6e6e6;");
+        row.setOnMouseClicked(e -> loadThread(q));
+        row.setOnMouseEntered(e -> row.setStyle("-fx-background-color:#f7faff; -fx-border-color:#cfe3ff;"));
+        row.setOnMouseExited(e -> row.setStyle("-fx-background-color:#ffffff; -fx-border-color:#e6e6e6;"));
+        return row;
+    }
+
+    private void loadThread(Question q) {
+        displayQuestionThread.getChildren().clear();
+
+        // header row with action buttons
+        HBox headerActions = new HBox(10);
+        Button logBtn = new Button("View Activity Log");
+        logBtn.setOnAction(e -> openLogDialog(q));
+        headerActions.getChildren().add(logBtn);
+
+        boolean isAsker = currentUser.getUserName() != null &&
+                currentUser.getUserName().equalsIgnoreCase(q.getAskedBy());
+        boolean isStaff = isStaff(currentUser.getRole());
+
+        if (isAsker || isStaff) {
+            Button followup = new Button("Create revised question");
+            followup.setOnAction(e -> openCreateFollowupDialog(q));
+            headerActions.getChildren().add(followup);
+        }
+        
+        // Add edit/delete buttons for question owner
+        if (isAsker && !q.isResolved()) {
+            Button editBtn = new Button("Edit Question");
+            editBtn.setStyle("-fx-background-color: #ff9800; -fx-text-fill: white;");
+            editBtn.setOnAction(e -> {
+                editQuestion(q);
+                refreshQuestionListAndMaybeSelect();
+            });
+            headerActions.getChildren().add(editBtn);
+        }
+        
+        if (isAsker) {
+            Button deleteBtn = new Button("Delete Question");
+            deleteBtn.setStyle("-fx-background-color: #f44336; -fx-text-fill: white;");
+            deleteBtn.setOnAction(e -> {
+                deleteQuestion(q);
+                refreshQuestionListAndMaybeSelect();
+            });
+            headerActions.getChildren().add(deleteBtn);
+        }
+
+        // header (question)
+        Label title = new Label(q.getTitle());
+        title.setStyle("-fx-font-size:18px; -fx-font-weight:bold; -fx-text-fill: black;");
+        Label meta = new Label("by " + q.getAskedBy() + " • " + q.getCreatedAt().format(TS));
+        meta.setStyle("-fx-text-fill: black; -fx-font-size:12px;");
+        Label body = new Label(q.getContent());
+        body.setWrapText(true);
+        body.setStyle("-fx-padding:8; -fx-background-color:#fff; -fx-border-color:#ddd; -fx-text-fill: black;");
+
+        VBox header = new VBox(6, headerActions, title, meta, body);
+        header.setPadding(new Insets(6,0,12,0));
+        displayQuestionThread.getChildren().add(header);
+
+        // If asker opened the thread, auto-mark unread private messages
+        if (isAsker) {
+            try { databaseHelper.markPrivateMessagesReadByAsker(q.getId(), currentUser.getUserName()); }
+            catch (SQLException ignored) {}
+        }
+
+        // Thread body based on filter
+        String mode = viewFilter.getValue();
+        if ("Private".equals(mode)) {
+            renderPrivateSection(q);
+        } else if ("All".equals(mode)) {
+            renderPublicAnswers(q);
+            renderPrivateSection(q);
+        } else { // Public
+            renderPublicAnswers(q);
+        }
+
+        // revised children list
+        renderFollowups(q);
+
+        // composer(s)
+        renderComposers(q, mode);
+
+        // react to filter changes
+        viewFilter.setOnAction(ev -> loadThread(q));
+    }
+
+    private void renderPublicAnswers(Question q) {
+        try {
+            List<Answer> answers = databaseHelper.getAnswersForQuestion(q.getId());
+            
+            // Header with "View All Answers" button
+            HBox ansHeader = new HBox(10);
+            ansHeader.setAlignment(Pos.CENTER_LEFT);
+            
+            Label ansTitle = new Label("Answers (" + answers.size() + ")");
+            ansTitle.setStyle("-fx-font-weight:bold; -fx-text-fill: black;");
+            
+            Button viewAllBtn = new Button("View All Answers Dialog");
+            viewAllBtn.setStyle("-fx-background-color: #0099ff; -fx-text-fill: white; -fx-font-size: 11px;");
+            viewAllBtn.setOnAction(e -> {
+                // Reload question with fresh answer data
+                try {
+                    Question freshQ = databaseHelper.getQuestionById(q.getId());
+                    if (freshQ != null) {
+                        showAnswersDialog(freshQ);
+                    }
+                } catch (SQLException ex) {
+                    showError("Load failed", ex.getMessage());
+                }
+            });
+            
+            ansHeader.getChildren().addAll(ansTitle, viewAllBtn);
+            VBox answersBox = new VBox(8, ansHeader);
+
+            boolean isAsker = currentUser.getUserName() != null &&
+                    currentUser.getUserName().equalsIgnoreCase(q.getAskedBy());
+
+            if (answers.isEmpty()) {
+                Label none = new Label("No answers yet. Be the first to help!");
+                none.setStyle("-fx-text-fill: black;");
+                answersBox.getChildren().add(none);
+            } else {
+                for (Answer a : answers) {
+                    VBox card = new VBox(5);
+                    
+                    // Answer metadata and content
+                    Label aMeta = new Label(a.getAnsweredBy() + " • " + a.getCreatedAt().format(TS) + " • " + a.getUpvotes() + " upvotes");
+                    aMeta.setStyle("-fx-text-fill: black; -fx-font-size:11px;");
+                    Label aText = new Label(a.getContent());
+                    aText.setWrapText(true);
+                    aText.setStyle("-fx-text-fill: black;");
+                    
+                    // Action buttons
+                    HBox actionBox = new HBox(5);
+                    
+                    // Upvote button (for everyone except the answer author)
+                    if (!a.getAnsweredBy().equals(currentUser.getUserName())) {
+                        Button upvoteBtn = new Button("👍 Upvote");
+                        upvoteBtn.setStyle("-fx-font-size: 10px;");
+                        upvoteBtn.setOnAction(e -> {
+                            try {
+                                databaseHelper.upvoteAnswer(a.getId(), a.getAnsweredBy());
+                                showAlert("Success", "Upvoted!", AlertType.INFORMATION);
+                                loadThread(q); // Refresh
+                            } catch (SQLException ex) {
+                                showAlert("Error", "Failed to upvote: " + ex.getMessage(), AlertType.ERROR);
+                            }
+                        });
+                        actionBox.getChildren().add(upvoteBtn);
+                    }
+                    
+                    // Mark as solution button (only for question owner on unresolved questions)
+                    if (isAsker && !q.isResolved()) {
+                        Button markBtn = new Button("✓ Mark as Solution");
+                        markBtn.setStyle("-fx-background-color: #4CAF50; -fx-text-fill: white; -fx-font-size: 10px;");
+                        markBtn.setOnAction(e -> {
+                            try {
+                                databaseHelper.markQuestionResolved(q.getId(), a.getId(), currentUser.getUserName());
+                                showAlert("Success", "Question marked as resolved!", AlertType.INFORMATION);
+                                refreshQuestionListAndMaybeSelect();
+                            } catch (SQLException ex) {
+                                showAlert("Error", "Failed to mark as resolved: " + ex.getMessage(), AlertType.ERROR);
+                            }
+                        });
+                        actionBox.getChildren().add(markBtn);
+                    }
+                    
+                    // Edit/Delete buttons (only for answer owner)
+                    if (a.getAnsweredBy().equals(currentUser.getUserName())) {
+                        Button editBtn = new Button("Edit");
+                        editBtn.setStyle("-fx-font-size: 10px;");
+                        editBtn.setOnAction(e -> {
+                            editAnswer(a);
+                            loadThread(q); // Refresh
+                        });
+                        
+                        Button deleteBtn = new Button("Delete");
+                        deleteBtn.setStyle("-fx-background-color: #f44336; -fx-text-fill: white; -fx-font-size: 10px;");
+                        deleteBtn.setOnAction(e -> {
+                            try {
+                                databaseHelper.deleteAnswer(a.getId(), currentUser.getUserName());
+                                showAlert("Success", "Answer deleted!", AlertType.INFORMATION);
+                                loadThread(q); // Refresh
+                            } catch (SQLException ex) {
+                                showAlert("Error", "Failed to delete answer: " + ex.getMessage(), AlertType.ERROR);
+                            }
+                        });
+                        
+                        actionBox.getChildren().addAll(editBtn, deleteBtn);
+                    }
+                    
+                    card.getChildren().addAll(aMeta, aText, actionBox);
+                    
+                    // Highlight if this is the accepted answer
+                    if (q.isResolved() && a.getId() == q.getResolvedAnswerId()) {
+                        card.setStyle("-fx-background-color:#e8f5e9; -fx-border-color:#4CAF50; -fx-border-width:2; -fx-padding:8;");
+                        Label acceptedLabel = new Label("✓ Accepted Answer");
+                        acceptedLabel.setStyle("-fx-font-weight: bold; -fx-text-fill: #2e7d32; -fx-font-size:10px;");
+                        card.getChildren().add(0, acceptedLabel);
+                    } else {
+                        card.setStyle("-fx-background-color:#f9f9f9; -fx-border-color:#eee; -fx-padding:8;");
+                    }
+                    
+                    answersBox.getChildren().add(card);
+                }
+            }
+            displayQuestionThread.getChildren().add(answersBox);
+        } catch (SQLException ex) {
+            showError("Answers load failed", ex.getMessage());
+        }
+    }
+
+    private void renderPrivateSection(Question q) {
+        VBox box = new VBox(8);
+        
+        try {
+            List<PrivateMessage> all = databaseHelper.getPrivateMessagesForQuestion(q.getId());
+            
+            // Count visible messages first
+            int visibleCount = 0;
+            for (PrivateMessage pm : all) {
+                if (canSeePrivate(q, pm)) {
+                    visibleCount++;
+                }
+            }
+            
+            // Create the title with the count
+            Label title = new Label("Private feedback (" + visibleCount + ")");
+            title.setStyle("-fx-font-weight:bold; -fx-text-fill: black;");
+            box.getChildren().add(title);
+            
+            // Render the messages
+            boolean any = false;
+            for (PrivateMessage pm : all) {
+                if (!canSeePrivate(q, pm)) continue;
+                any = true;
+                VBox card = new VBox(3);
+                Label meta = new Label(pm.getSender() + " • " + pm.getCreatedAt().format(TS) + " • " + pm.getMessageType());
+                meta.setStyle("-fx-text-fill: black; -fx-font-size:11px;");
+                Label text = new Label(pm.getContent());
+                text.setWrapText(true);
+                text.setStyle("-fx-text-fill: black;");
+                card.getChildren().addAll(meta, text);
+                card.setStyle("-fx-background-color:#fffbea; -fx-border-color:#ffe08a; -fx-padding:8;");
+                box.getChildren().add(card);
+            }
+            
+            if (!any) {
+                Label none = new Label("No private feedback yet.");
+                none.setStyle("-fx-text-fill: black;");
+                box.getChildren().add(none);
+            }
+        } catch (SQLException ex) {
+            showError("Private messages load failed", ex.getMessage());
+        }
+
+        displayQuestionThread.getChildren().add(box);
+    }
+
+    private void renderFollowups(Question q) {
+        try {
+            List<Question> kids = databaseHelper.getFollowupQuestions(q.getId());
+            if (kids == null || kids.isEmpty()) return;
+            Label head = new Label("Revised questions");
+            head.setStyle("-fx-font-weight:bold;");
+            VBox list = new VBox(6);
+            for (Question k : kids) {
+                Hyperlink link = new Hyperlink(k.getTitle() + " — " + k.getCreatedAt().format(TS));
+                link.setOnAction(e -> loadThread(k));
+                list.getChildren().add(link);
+            }
+            VBox wrap = new VBox(6, head, list);
+            wrap.setPadding(new Insets(8,0,0,0));
+            displayQuestionThread.getChildren().add(wrap);
+        } catch (SQLException ex) {
+            showError("Followups load failed", ex.getMessage());
+        }
+    }
+
+    private void renderComposers(Question q, String mode) {
+        // public composer
+        if (!"Private".equals(mode)) {
+            displayQuestionThread.getChildren().add(answerComposer(q));
+        }
+        // private composer
+        if (!"Public".equals(mode)) {
+            displayQuestionThread.getChildren().add(privateComposer(q));
+        }
+    }
+
+    private VBox answerComposer(Question q) {
+        Label lbl = new Label("Add a public answer");
+        lbl.setStyle("-fx-font-weight:bold; -fx-text-fill: black;");
+        TextArea text = new TextArea();
+        text.setPromptText("Write your answer...");
+        text.setWrapText(true);
+        text.setPrefRowCount(3);
+        Button post = new Button("Post Public Answer");
+        post.setOnAction(e -> {
+            String content = trimOrEmpty(text.getText());
+            if (content.length() < 5) {
+                showError("Too short", "Please write a longer answer.");
+                return;
+            }
+            try {
+                Answer a = new Answer(q.getId(), content, currentUser.getUserName());
+                databaseHelper.createAnswer(a);
+                loadThread(q); // refresh thread
+            } catch (SQLException ex) {
+                showError("Post failed", ex.getMessage());
+            }
+        });
+        VBox box = new VBox(6, lbl, text, post);
+        box.setPadding(new Insets(10,0,0,0));
+        return box;
+    }
+
+    private VBox privateComposer(Question q) {
+        Label lbl = new Label("Send private feedback to the asker");
+        lbl.setStyle("-fx-font-weight:bold; -fx-text-fill: black;");
+        ComboBox<String> kind = new ComboBox<>();
+        kind.getItems().addAll("Question", "Answer");
+        kind.setValue("Question");
+        TextArea text = new TextArea();
+        text.setPromptText("Write your private message...");
+        text.setWrapText(true);
+        text.setPrefRowCount(3);
+        Button send = new Button("Send Private");
+        send.setOnAction(e -> {
+            String content = trimOrEmpty(text.getText());
+            if (content.length() < 5) {
+                showError("Too short", "Please write a longer message.");
+                return;
+            }
+            try {
+                databaseHelper.addPrivateMessage(q.getId(), currentUser.getUserName(), kind.getValue().toUpperCase(), content);
+                loadThread(q);
+            } catch (SQLException ex) {
+                showError("Send failed", ex.getMessage());
+            }
+        });
+        VBox box = new VBox(6, lbl, new HBox(6, new Label("Type:") {{ setStyle("-fx-text-fill: black;"); }}, kind), text, send);
+        box.setPadding(new Insets(10,0,0,0));
+        return box;
+    }
+
+    private void openCreateQuestionDialog() {
+        Dialog<ButtonType> dialog = new Dialog<>();
+        dialog.setTitle("New Question");
+
+        Label l1 = new Label("Title:");
+        TextField tfTitle = new TextField();
+        tfTitle.setPromptText("Short summary of your problem");
+
+        Label l2 = new Label("Content:");
+        TextArea taContent = new TextArea();
+        taContent.setPromptText("Describe the problem in more detail");
+        taContent.setWrapText(true);
+        taContent.setPrefRowCount(4);
+
+        GridPane gp = new GridPane();
+        gp.setHgap(8); gp.setVgap(8); gp.setPadding(new Insets(10));
+        gp.add(l1, 0, 0); gp.add(tfTitle, 1, 0);
+        gp.add(l2, 0, 1); gp.add(taContent, 1, 1);
+
+        dialog.getDialogPane().setContent(gp);
+        dialog.getDialogPane().getButtonTypes().addAll(ButtonType.CANCEL, ButtonType.OK);
+
+        dialog.setResultConverter(btn -> btn);
+
+        dialog.showAndWait().ifPresent(btn -> {
+            if (btn == ButtonType.OK) {
+                String t = trimOrEmpty(tfTitle.getText());
+                String c = trimOrEmpty(taContent.getText());
+                if (t.length() < 5 || c.length() < 10) {
+                    showError("Invalid input", "Please provide a longer title and content.");
+                    return;
+                }
+                try {
+                    Question q = new Question(t, c, currentUser.getUserName());
+                    databaseHelper.createQuestion(q);
+                    refreshQuestionListAndMaybeSelect();
+                } catch (SQLException ex) {
+                    showError("Create failed", ex.getMessage());
+                }
+            }
+        });
+    }
+
+    private void openCreateFollowupDialog(Question parent) {
+        Dialog<ButtonType> dialog = new Dialog<>();
+        dialog.setTitle("Create revised question");
+
+        Label l1 = new Label("Title:");
+        TextField tfTitle = new TextField();
+        tfTitle.setPromptText("Updated title based on feedback");
+
+        Label l2 = new Label("Content:");
+        TextArea taContent = new TextArea();
+        taContent.setPromptText("Write the revised question content");
+        taContent.setWrapText(true);
+        taContent.setPrefRowCount(4);
+
+        GridPane gp = new GridPane();
+        gp.setHgap(8); gp.setVgap(8); gp.setPadding(new Insets(10));
+        gp.add(l1, 0, 0); gp.add(tfTitle, 1, 0);
+        gp.add(l2, 0, 1); gp.add(taContent, 1, 1);
+
+        dialog.getDialogPane().setContent(gp);
+        dialog.getDialogPane().getButtonTypes().addAll(ButtonType.CANCEL, ButtonType.OK);
+
+        dialog.showAndWait().ifPresent(btn -> {
+            if (btn == ButtonType.OK) {
+                String t = trimOrEmpty(tfTitle.getText());
+                String c = trimOrEmpty(taContent.getText());
+                if (t.length() < 5 || c.length() < 10) {
+                    showError("Invalid input", "Please provide a longer title and content.");
+                    return;
+                }
+                try {
+                    databaseHelper.createFollowupQuestion(parent.getId(), t, c, currentUser.getUserName());
+                    loadThread(parent); // reload; children list will refresh
+                } catch (SQLException ex) {
+                    showError("Create failed", ex.getMessage());
+                }
+            }
+        });
+    }
+
+    private void openLogDialog(Question q) {
+        Dialog<Void> dialog = new Dialog<>();
+        dialog.setTitle("Activity Log");
+        dialog.getDialogPane().getButtonTypes().addAll(ButtonType.CLOSE);
+
+        VBox content = new VBox(8);
+        content.setPadding(new Insets(10));
+
+        List<EventRecord> events = new ArrayList<>();
+        // question created
+        events.add(new EventRecord(q.getCreatedAt(), "Question created by " + q.getAskedBy() + ": " + q.getTitle()));
+
+        try {
+            // answers
+            for (Answer a : databaseHelper.getAnswersForQuestion(q.getId())) {
+                events.add(new EventRecord(a.getCreatedAt(), "Answer by " + a.getAnsweredBy() + ": " + a.getContent()));
+            }
+            // private (filter by visibility)
+            for (PrivateMessage pm : databaseHelper.getPrivateMessagesForQuestion(q.getId())) {
+                if (canSeePrivate(q, pm)) {
+                    events.add(new EventRecord(pm.getCreatedAt(), "Private " + pm.getMessageType() + " by " + pm.getSender() + ": " + pm.getContent()));
+                }
+            }
+            // followups
+            for (Question child : databaseHelper.getFollowupQuestions(q.getId())) {
+                events.add(new EventRecord(child.getCreatedAt(), "Revised question created by " + child.getAskedBy() + ": " + child.getTitle()));
+            }
+        } catch (SQLException ex) {
+            content.getChildren().add(new Label("Failed to load log: " + ex.getMessage()));
+        }
+
+        events.sort(Comparator.comparing(e -> e.t));
+        for (EventRecord e : events) {
+            Label line = new Label("• " + TS.format(e.t) + " — " + e.text);
+            line.setWrapText(true);
+            content.getChildren().add(line);
+        }
+
+        ScrollPane sp = new ScrollPane(content);
+        sp.setFitToWidth(true);
+        dialog.getDialogPane().setContent(sp);
+        dialog.showAndWait();
+    }
+
+    // --- utils ---
+
+    private boolean canSeePrivate(Question q, PrivateMessage pm) {
+        boolean isAsker = currentUser.getUserName() != null &&
+                currentUser.getUserName().equalsIgnoreCase(q.getAskedBy());
+        boolean isSender = currentUser.getUserName() != null &&
+                currentUser.getUserName().equalsIgnoreCase(pm.getSender());
+        boolean isStaff = isStaff(currentUser.getRole());
+        return isAsker || isSender || isStaff;
+    }
+
+    private boolean isStaff(String role) {
+        if (role == null) return false;
+        String r = role.toLowerCase();
+        return r.contains("admin") || r.contains("instructor") || r.contains("staff");
+    }
+
+    private String trimOrEmpty(String s) {
+        return (s == null) ? "" : s.trim();
+    }
+    
+    // Simple error alert helper used by posting/loading actions.
+    private void showError(String title, String message) {
+        Alert alert = new Alert(Alert.AlertType.ERROR, message, ButtonType.OK);
+        alert.setTitle(title);
+        alert.setHeaderText(null); // keep it clean
+        alert.showAndWait();
+    }  
+
+    // simple record for the log
+    private static class EventRecord {
+        LocalDateTime t;
+        String text;
+        EventRecord(LocalDateTime t, String text) { this.t = t; this.text = text; }
     }
 
     // Display all questions
@@ -662,18 +1322,21 @@ public class StudentQAPage {
         // Header
         HBox headerBox = new HBox(10);
         headerBox.setAlignment(Pos.CENTER_LEFT);
+
         Label titleLabel = new Label(question.getTitle());
-        titleLabel.setStyle("-fx-font-size: 16px; -fx-font-weight: bold; -fx-text-fill: #2c2c2c;");
+        titleLabel.setStyle("-fx-font-size:18px; -fx-font-weight:bold; -fx-text-fill:#2c2c2c;");
+
         Label statusLabel = new Label(question.isResolved() ? "✓ RESOLVED" : "");
-        statusLabel.setStyle("-fx-background-color: #4CAF50; -fx-text-fill: white; -fx-padding: 3 8; -fx-font-size: 10px;");
+        statusLabel.setStyle("-fx-background-color:#4CAF50; -fx-text-fill:white; -fx-padding:3 8; -fx-font-size:10px;");
         statusLabel.setVisible(question.isResolved());
+
         headerBox.getChildren().addAll(titleLabel, statusLabel);
 
 		
         // Content
         Label contentLabel = new Label(question.getContent());
         contentLabel.setWrapText(true);
-        contentLabel.setStyle("-fx-text-fill: #333;");
+        contentLabel.setStyle("-fx-padding:8; -fx-background-color:#fff; -fx-border-color:#ddd;");
 
 		
         // Metadata
